backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
pytesseract==0.3.10
Pillow==10.2.0
pdf2image==1.17.0
supabase==2.3.4
python-dotenv==1.0.1
pydantic==2.6.1
qrcode[pil]==7.4.2
httpx==0.27.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
regex==2024.1.1
